% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_indices.R
\name{calc_ndvi}
\alias{calc_ndvi}
\title{Calcul du NDVI (Normalized Difference Vegetation Index)}
\usage{
calc_ndvi(nir, red)
}
\arguments{
\item{nir}{Bande NIR : valeur numérique, vecteur ou SpatRaster.}

\item{red}{Bande Rouge : valeur numérique, vecteur ou SpatRaster.}
}
\value{
Même type que l'entrée (scalaire, vecteur ou SpatRaster).
Les valeurs sont comprises entre -1 et 1.
}
\description{
Calcule le NDVI à partir des bandes Rouge et NIR.
Formule : NDVI = (NIR - Red) / (NIR + Red).
Fonctionne sur des scalaires, des vecteurs numériques ou des SpatRasters.
}
\examples{
# Scalaire
calc_ndvi(nir = 0.8, red = 0.1)
# [1] 0.7777778

# Vecteur
calc_ndvi(nir = c(0.7, 0.6, 0.5), red = c(0.1, 0.2, 0.3))

\dontrun{
# SpatRaster
nir_r  <- terra::rast("nir_band.tif")
red_r  <- terra::rast("red_band.tif")
ndvi_r <- calc_ndvi(nir = nir_r, red = red_r)
terra::plot(ndvi_r, main = "NDVI")
}

}
