% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_sentinel.R
\name{.load_sentinel_local}
\alias{.load_sentinel_local}
\title{Charge des bandes Sentinel-2 depuis un répertoire local}
\usage{
.load_sentinel_local(sentinel_dir, out_dir)
}
\description{
Charge des bandes Sentinel-2 depuis un répertoire local
}
\keyword{internal}
