% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_sentinel.R
\name{.modis_to_raster}
\alias{.modis_to_raster}
\title{Convertit un data.frame MODISTools en SpatRaster multi-dates}
\usage{
.modis_to_raster(df, scale_factor = 1e-04)
}
\description{
Convertit un data.frame MODISTools en SpatRaster multi-dates
}
\keyword{internal}
